requirements.txt
# VaultRunner Dependencies

# Core dependencies
PyYAML>=6.0
requests>=2.25.0
cryptography>=3.4.0
aiohttp>=3.8.0

# Optional dependencies for enhanced functionality
docker>=5.0.0
kubernetes>=18.0.0

# Development dependencies (install with pip install -e .[dev])
# pytest>=6.0
# pytest-cov>=2.0
# black>=21.0
# flake8>=3.8
# mypy>=0.812